feat: Add Windows tool server and WSL agent dashboard scripts for mixed-environment setup
