+++ docs/WORKING_STATE.md
@@ -26,7 +26,7 @@ Purpose: current operational snapshot for any teammate or AI agent joining mid-h
 - `MODEL_PROVIDER=ollama` now defaults the same OpenAI-compatible client to `http://localhost:11434` with model `nemotron-3-nano:4b`, while `MODEL_PROVIDER=nim` preserves the prior NIM-oriented defaults and overrides.
 - The dashboard has a reset control wired through `POST /demo/reset-downloads` to restore the fake Downloads fixture for repeatable demos.
 - `scripts/run_local_stack.py` can bootstrap the local Python/dashboard dependencies and start the tool server, agent server, and dashboard from one terminal, skipping only `.env` creation.
-- `scripts/run_hybrid_stack.py` can launch the tool server in a separate interactive Windows PowerShell window while keeping the agent server and dashboard in WSL, matching the current mixed-environment desktop demo setup.
+- `scripts/run_windows_tool_server.py` can launch the tool server in a separate interactive Windows PowerShell window, and `scripts/run_wsl_agent_dashboard.py` can run the agent server plus dashboard in WSL, matching the current mixed-environment desktop demo setup.
 - In `TOOL_MODE=windows`, the desktop backend can launch allowlisted apps, create a fresh temp file for Notepad, focus the launched window when Windows allows it, and paste text through `keyboard_type`.
 - In `TOOL_MODE=windows`, `screen_capture` uses Pillow/ImageGrab to save a PNG screenshot under the local temp `NemotronOS/screenshots` directory and returns the saved file path plus dimensions.
 - The coordinator auto-follows known Notepad typing goals with `keyboard_type` after `app_launch`. It now prefers locally extracted quoted/trailing text from the user goal, then falls back to stored voice dictation text, and only asks the model for a second planning step when the literal text is still unclear.
